requirements.txt
acres==0.5.0
fastapi==0.115.6
uvicorn==0.32.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
audioread==3.0.1
banks==2.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
datasketch==1.6.5
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.30
lazy_loader==0.4
legacy==0.1.7
librosa==0.10.2.post1
llama-index-core==0.14.2
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-instrumentation==0.4.1
llama-index-legacy==0.9.48.post4
llama-index-vector-stores-pinecone==0.7.1
llama-index-workflows==2.2.2
llvmlite==0.45.0
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numba==0.62.0
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.108.2
orjson==3.11.3
packaging==24.2
pandas==2.3.2
pathlib==1.0.1
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
prov==2.1.1
puremagic==1.30
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
Resemblyzer==0.1.4
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.3.1+cpu
tqdm==4.67.1
traits==7.0.2
transformers==4.56.2
triton==3.4.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webrtcvad==2.0.10
wrapt==1.17.3
yarl==1.20.1
pytest==8.3.4
google-api-python-client==2.129.0
google-cloud-storage==2.18.2
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
isodate==0.6.1
zstandard==0.25.0
